Generified the statistics modification script so that it takes a sed file as a parameter.

+++ kaiko/misc/fix_all_stats.sh
@@ -1,16 +1,21 @@
 #!/bin/bash
 
-SEDSCRIPT="~/develop/dbnary/kaiko/fix_stats.sed"
+SEDSCRIPT="$HOME/develop/dbnary/kaiko/misc/$1"
+if [[ ! -f $SEDSCRIPT ]] ; then
+	>&2 echo "usage : fix_all_stats.sh sed-file-name"
+	exit 1
+fi
+
 SAVEDIR="$HOME/stats.saved/`date +%F_%H-%M-%S`/"
 DATADIR="$HOME/develop/wiktionary/extracts/ontolex"
 mkdir -p $SAVEDIR
-rsync -a $DATADIR/ $SAVEDIR/
+rsync --include='??' --include='??/??_dbnary_stat*.ttl.bz2' --exclude='*.ttl*' -a $DATADIR/ $SAVEDIR/
 
 pushd $DATADIR
-for lg in ?? do
+for lg in ??; do
 	pushd $lg
 	# iterate over all dates...
-	for f in ${lg}_dbanry_stat*.ttl.bz2 do
+	for f in ${lg}_dbnary_stat*.ttl.bz2 ; do
 		bzcat $f | sed -f ${SEDSCRIPT} | bzip2 > $f.new
 		mv $f $f.old
 		mv $f.new $f
